refactor(ui-library): reorganize folder structure
