+++ packages/cube-frontend-ui-library/src/internal/components/StoryLayout/StoryLayout.tsx
