+++ packages/cube-frontend-ui-library/src/stories/designTokens/Typography/HeadingBox.tsx
